moving the rest of teachers function to teachers
